增加白名单

+++ 程序/ble_3435_sdk_ext_39_0D0D/sdk/plactform/rom/hci/BK_HCI_Protocol.h
@@ -34,10 +34,10 @@ typedef double              fp64;       // 双精度浮点数(64位长度)
 
 typedef signed   char       s8;
 typedef signed   short      s16;
-typedef signed   long       s32;
+typedef signed   int       s32;
 typedef unsigned char       u8;
 typedef unsigned short      u16;
-typedef unsigned long       u32;
+typedef unsigned int       u32;
 //typedef unsigned long       u_int32;
 
 typedef unsigned char  		UINT8;      /* Unsigned  8 bit quantity        */
